audit-agent/backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20       # File uploads

# ── Validation & Settings ─────────────────────────────────────────────────────
pydantic==2.10.6
pydantic-settings==2.7.1

# ── AI / LLM ──────────────────────────────────────────────────────────────────
openai==1.75.0                 # GPT-4o (OpenAI SDK — dev phase)

# ── PDF Processing ────────────────────────────────────────────────────────────
pymupdf==1.25.3                # PyMuPDF: fast native PDF text extraction
pymupdf4llm==0.0.17            # Markdown-aware PDF → LLM conversion
Pillow==11.1.0                 # Image manipulation for OCR pages

# ── PII Anonymization (local, GDPR-compliant) ────────────────────────────────
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.8.2

# ── Database ──────────────────────────────────────────────────────────────────
supabase==2.13.0
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.14.1

# ── Storage ───────────────────────────────────────────────────────────────────
boto3==1.37.10                 # S3-compatible (AWS eu-central-1)

# ── Queue / Tasks ─────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1
flower==2.0.1                  # Celery monitoring dashboard

# ── Auth & Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4
slowapi==0.1.9                    # Rate limiting

# ── Date/Time ─────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2025.1

# ── Observability ─────────────────────────────────────────────────────────────
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-exporter-otlp==1.30.0
prometheus-fastapi-instrumentator==7.1.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity==9.0.0                # Exponential retry
structlog==25.1.0              # Structured logging
python-ulid==3.0.0             # Sortable unique IDs

# ── Dev / Test ────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
httpx==0.28.1                  # TestClient async
ruff==0.9.7                    # Linter + formatter
mypy==1.14.1                   # Static type checker
bandit==1.8.3                  # Security scanner
